+++ 013/cellular.js
@@ -9,7 +9,7 @@ module.exports = function (regl) {
 
   return regl({
     vert: glsl`
-      precision mediump float;
+      precision highp float;
       attribute vec3 normal, position;
       uniform mat4 model, view, projection;
       varying vec3 vNormal;
@@ -20,7 +20,7 @@ module.exports = function (regl) {
       }
     `,
     frag: glsl`
-      precision mediump float;
+      precision highp float;
       #pragma glslify: matcap = require(matcap)
       uniform vec3 cameraPosition;
       uniform sampler2D matcapTexture;
